+++ Changelog.md
@@ -8,7 +8,7 @@ Added `ItemAdapter.is_item_class` and `ItemAdapter.get_field_meta_from_class`
 
 ### 0.3.0 (2021-07-15)
 
-Added suport for `pydantic` models ([#53](https://github.com/scrapy/itemadapter/pull/53))
+Added built-in support for `pydantic` models ([#53](https://github.com/scrapy/itemadapter/pull/53))
 
 
 ### 0.2.0 (2020-11-06)